Fix an issue where util.inspect was throwing an error
With the latest lodash v4.7 upgrade, Symbol keys are included in lodash's _.keysIn array. However, due to a bug in harmony-reflect, calling Object.getOwnPropertySymbols() on a shimmed Proxy causes an error to be thrown. Until this bug gets fixed, it's best to avoid anything that calls that function on a proxied object. Since _.omitBy calls _.keysIn internally, it was causing a bug when used on a Proxy, so it was replaced with _.pick which does not have this issue.