Rename the _ac internal property to _r
It seems like 'r' that's the canonical name for a requester at this point. _ac is actually a remnant from when the project was brand-new, and the requester class was called 'AuthenticatedClient', but now it's sort of misleading and anyone unfamiliar with the code probably has no idea what it is.