Don't throw a 404 error when noop-unsubscribing from a subreddit.
The reddit API returns a 404 error if the user attempts to unsubscribe from a subreddit to which they were never subscribed in the first place. This is a bit misleading (especially since no analogous error gets thrown when double-subscribing), so it's best to catch the error internally.

One complication, however, is that there is no observable difference between the 404 error that occurs as a result of unsubscribing, and the 404 error that occurs when the subreddit itself does not exist. For instance, r.get_subreddit('afdafasfsfaesr').unsubscribe() *should* throw a 404 error because that subreddit does not exist. On the other hand, r.get_subreddit('gifs').unsubscribe() should never reflect a 404 error, because that subreddit exists. To differentiate between these two cases, after receiving a 404 error on unsubscribe(), snoowrap attempts to fetch the info page for the subreddit in question. If the info page is retrieved successfully, snoowrap swallows the original 404 error because it is inapplicable.