Fix an issue with fetching threads containing many top-level comments
When reddit's api/morechildren returns a response after fetching 20 comments (the maximum number that can be fetched from this endpoint), it appends a 'more' object to the end of the list. This object was unaccounted for, so snoowrap was keeping it in the listing as if it was a regular Comment.