Fix the tests that were causing OOM errors
There was previously a bug in functions such as Subreddit#get_banned_users where object parameters were not getting passed as querystring parameters. This bug was fixed a few commits ago.

There was also a separate issue in a few of the tests, in which RedditUser objects were being passed as parameters to these functions, with the expectation that the  property would get stripped off, and all the other properties would be dropped because they weren't needed. However, when the first bug was fixed, request-promise attepted to send all of the properties of the unfetched RedditUser object as querystring parameters. RedditUser objects have a bunch of prototype functions, which might have gotten called as well. The end result was that a bunch of functions got called that shouldn't have been, and so an out-of-memory error occured.

The fix for this was to simply change the test case to avoid passing the entire RedditUser object as a parameter. However, it might also be a good idea to have a safeguard in the function itself to pick only the needed query parameters (i.e. the Listing parameters `before`, `after`, `limit`, etc. as well as the `name` parameter which is actually needed).