Use @SpyTec's map to loop through all of
values and check if they are a submission or comment object, if so, grab the fullname ID, if it's just a string, check if it's valid, if it isn't throw a error, if it is return. If it's none of the above, throw an error. Also had to add in a check to ensure we were getting arrays we could map, if we weren't, it would error out in a unpleasant way. This also adds in two examples of how to use this function.